♻️ getRoomList에 slice 로직 추가
