+++ src/services/lobby.js
@@ -1,6 +1,5 @@
 import { getRoomList } from '@/services/game';
 import { getUserProfile } from '@/services/user';
-import { sliceRoomList } from '@/util/sliceRoomList';
 
 export const lobbyLoader = async () => {
   const [roomList, userProfile] = await Promise.all([
@@ -8,7 +7,5 @@ export const lobbyLoader = async () => {
     getUserProfile(),
   ]);
 
-  const slicedRoomList = sliceRoomList({ roomList });
-
-  return { roomList: slicedRoomList, userProfile };
+  return { roomList, userProfile };
 };
